Rework server resource
Three main things are reworked in that patch. Firstly, the update of
the hostname and IP are moved in the Update handler. Then the apply changes
is computed by terraform core, so it is useless to perform the check with
`updateServerIfNeeded`. Finally, the init of the Server struct from
the Server endpoint is also removed.

Signed-off-by: Nicolas Gelot <nicofonk@gmail.com>